Adds stake.remove - working
