+++ bittensor_cli/src/commands/stake/stake.py
@@ -4,8 +4,7 @@ from typing import TYPE_CHECKING, Optional
 import typer
 
 from bittensor_wallet import Wallet
-from bittensor_wallet.errors import KeyFileError
-from rich.prompt import Confirm, Prompt
+from rich.prompt import Prompt
 from rich.table import Table
 from rich import box
 from rich.progress import Progress, BarColumn, TextColumn
@@ -18,816 +17,13 @@ from bittensor_cli.src.bittensor.chain_data import StakeInfo
 from bittensor_cli.src.bittensor.utils import (
     # TODO add back in caching
     console,
-    err_console,
-    print_verbose,
     print_error,
-    get_hotkey_wallets_for_wallet,
-    is_valid_ss58_address,
-    format_error_message,
-    group_subnets,
     millify_tao,
     get_subnet_name,
 )
 
 if TYPE_CHECKING:
     from bittensor_cli.src.bittensor.subtensor_interface import SubtensorInterface
-
-
-async def unstake_selection(
-    subtensor: "SubtensorInterface",
-    wallet: Wallet,
-    dynamic_info,
-    identities,
-    old_identities,
-    netuid: Optional[int] = None,
-):
-    stake_infos = await subtensor.get_stake_for_coldkey(
-        coldkey_ss58=wallet.coldkeypub.ss58_address
-    )
-
-    if not stake_infos:
-        print_error("You have no stakes to unstake.")
-        raise typer.Exit()
-
-    hotkey_stakes = {}
-    for stake_info in stake_infos:
-        if netuid is not None and stake_info.netuid != netuid:
-            continue
-        hotkey_ss58 = stake_info.hotkey_ss58
-        netuid_ = stake_info.netuid
-        stake_amount = stake_info.stake
-        if stake_amount.tao > 0:
-            hotkey_stakes.setdefault(hotkey_ss58, {})[netuid_] = stake_amount
-
-    if not hotkey_stakes:
-        if netuid is not None:
-            print_error(f"You have no stakes to unstake in subnet {netuid}.")
-        else:
-            print_error("You have no stakes to unstake.")
-        raise typer.Exit()
-
-    hotkeys_info = []
-    for idx, (hotkey_ss58, netuid_stakes) in enumerate(hotkey_stakes.items()):
-        if hk_identity := identities["hotkeys"].get(hotkey_ss58):
-            hotkey_name = hk_identity.get("identity", {}).get(
-                "name", ""
-            ) or hk_identity.get("display", "~")
-        elif old_identity := old_identities.get(hotkey_ss58):
-            hotkey_name = old_identity.display
-        else:
-            hotkey_name = "~"
-        # TODO: Add wallet ids here.
-
-        hotkeys_info.append(
-            {
-                "index": idx,
-                "identity": hotkey_name,
-                "netuids": list(netuid_stakes.keys()),
-                "hotkey_ss58": hotkey_ss58,
-            }
-        )
-
-    # Display existing hotkeys, id, and staked netuids.
-    subnet_filter = f" for Subnet {netuid}" if netuid is not None else ""
-    table = Table(
-        title=f"\n[{COLOR_PALETTE['GENERAL']['HEADER']}]Hotkeys with Stakes{subnet_filter}\n",
-        show_footer=True,
-        show_edge=False,
-        header_style="bold white",
-        border_style="bright_black",
-        style="bold",
-        title_justify="center",
-        show_lines=False,
-        pad_edge=True,
-    )
-    table.add_column("Index", justify="right")
-    table.add_column("Identity", style=COLOR_PALETTE["GENERAL"]["SUBHEADING"])
-    table.add_column("Netuids", style=COLOR_PALETTE["GENERAL"]["NETUID"])
-    table.add_column("Hotkey Address", style=COLOR_PALETTE["GENERAL"]["HOTKEY"])
-
-    for hotkey_info in hotkeys_info:
-        index = str(hotkey_info["index"])
-        identity = hotkey_info["identity"]
-        netuids = group_subnets([n for n in hotkey_info["netuids"]])
-        hotkey_ss58 = hotkey_info["hotkey_ss58"]
-        table.add_row(index, identity, netuids, hotkey_ss58)
-
-    console.print("\n", table)
-
-    # Prompt to select hotkey to unstake.
-    hotkey_options = [str(hotkey_info["index"]) for hotkey_info in hotkeys_info]
-    hotkey_idx = Prompt.ask(
-        "\nEnter the index of the hotkey you want to unstake from",
-        choices=hotkey_options,
-    )
-    selected_hotkey_info = hotkeys_info[int(hotkey_idx)]
-    selected_hotkey_ss58 = selected_hotkey_info["hotkey_ss58"]
-    selected_hotkey_name = selected_hotkey_info["identity"]
-    netuid_stakes = hotkey_stakes[selected_hotkey_ss58]
-
-    # Display hotkey's staked netuids with amount.
-    table = Table(
-        title=f"\n[{COLOR_PALETTE['GENERAL']['HEADER']}]Stakes for hotkey \n[{COLOR_PALETTE['GENERAL']['SUBHEADING']}]{selected_hotkey_name}\n{selected_hotkey_ss58}\n",
-        show_footer=True,
-        show_edge=False,
-        header_style="bold white",
-        border_style="bright_black",
-        style="bold",
-        title_justify="center",
-        show_lines=False,
-        pad_edge=True,
-    )
-    table.add_column("Subnet", justify="right")
-    table.add_column("Symbol", style=COLOR_PALETTE["GENERAL"]["SYMBOL"])
-    table.add_column("Stake Amount", style=COLOR_PALETTE["STAKE"]["STAKE_AMOUNT"])
-    table.add_column(
-        f"[bold white]RATE ({Balance.get_unit(0)}_in/{Balance.get_unit(1)}_in)",
-        style=COLOR_PALETTE["POOLS"]["RATE"],
-        justify="left",
-    )
-
-    for netuid_, stake_amount in netuid_stakes.items():
-        symbol = dynamic_info[netuid_].symbol
-        rate = f"{dynamic_info[netuid_].price.tao:.4f} τ/{symbol}"
-        table.add_row(str(netuid_), symbol, str(stake_amount), rate)
-    console.print("\n", table, "\n")
-
-    # Ask which netuids to unstake from for the selected hotkey.
-    unstake_all = False
-    if netuid is not None:
-        selected_netuids = [netuid]
-    else:
-        while True:
-            netuid_input = Prompt.ask(
-                "\nEnter the netuids of the [blue]subnets to unstake[/blue] from (comma-separated), or '[blue]all[/blue]' to unstake from all",
-                default="all",
-            )
-
-            if netuid_input.lower() == "all":
-                selected_netuids = list(netuid_stakes.keys())
-                unstake_all = True
-                break
-            else:
-                try:
-                    netuid_list = [int(n.strip()) for n in netuid_input.split(",")]
-                    invalid_netuids = [n for n in netuid_list if n not in netuid_stakes]
-                    if invalid_netuids:
-                        print_error(
-                            f"The following netuids are invalid or not available: {', '.join(map(str, invalid_netuids))}. Please try again."
-                        )
-                    else:
-                        selected_netuids = netuid_list
-                        break
-                except ValueError:
-                    print_error(
-                        "Please enter valid netuids (numbers), separated by commas, or 'all'."
-                    )
-
-    hotkeys_to_unstake_from = []
-    for netuid_ in selected_netuids:
-        hotkeys_to_unstake_from.append(
-            (selected_hotkey_name, selected_hotkey_ss58, netuid_)
-        )
-    return hotkeys_to_unstake_from, unstake_all
-
-
-def ask_unstake_amount(
-    current_stake_balance: Balance,
-    netuid: int,
-    staking_address_name: str,
-    staking_address_ss58: str,
-    interactive: bool,
-) -> Optional[Balance]:
-    """Prompt the user to decide the amount to unstake."""
-    while True:
-        response = Prompt.ask(
-            f"Unstake all: [{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]{current_stake_balance}[/{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]"
-            f" from [{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]{staking_address_name if staking_address_name else staking_address_ss58}[/{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]"
-            f" on netuid: [{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]{netuid}[/{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]? [y/n/q]",
-            choices=["y", "n", "q"],
-            default="n" if interactive else "y",
-            show_choices=True,
-        ).lower()
-
-        if response == "q":
-            return None  # Quit
-
-        elif response == "y":
-            return current_stake_balance
-
-        elif response == "n":
-            while True:
-                amount_input = Prompt.ask(
-                    f"Enter amount to unstake in [{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]{Balance.get_unit(netuid)}[/{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]"
-                    f" from subnet: [{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]{netuid}[/{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]"
-                    f" (Max: [{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]{current_stake_balance}[/{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}])"
-                )
-                if amount_input.lower() == "q":
-                    return None  # Quit
-
-                try:
-                    amount_value = float(amount_input)
-                    if amount_value <= 0:
-                        console.print("[red]Amount must be greater than zero.[/red]")
-                        continue  # Re-prompt
-
-                    amount_to_unstake = Balance.from_tao(amount_value)
-                    amount_to_unstake.set_unit(netuid)
-                    if amount_to_unstake > current_stake_balance:
-                        console.print(
-                            f"[red]Amount exceeds current stake balance of {current_stake_balance}.[/red]"
-                        )
-                        continue  # Re-prompt
-
-                    return amount_to_unstake
-
-                except ValueError:
-                    console.print(
-                        "[red]Invalid input. Please enter a numeric value or 'q' to quit.[/red]"
-                    )
-
-        else:
-            console.print("[red]Invalid input. Please enter 'y', 'n', or 'q'.[/red]")
-
-
-async def _unstake_all(
-    wallet: Wallet,
-    subtensor: "SubtensorInterface",
-    unstake_all_alpha: bool = False,
-    prompt: bool = True,
-) -> bool:
-    """Unstakes all stakes from all hotkeys in all subnets."""
-
-    with console.status(
-        f"Retrieving stake information & identities from {subtensor.network}...",
-        spinner="earth",
-    ):
-        (
-            stake_info,
-            ck_hk_identities,
-            old_identities,
-            all_sn_dynamic_info_,
-            current_wallet_balance,
-        ) = await asyncio.gather(
-            subtensor.get_stake_for_coldkey(wallet.coldkeypub.ss58_address),
-            subtensor.fetch_coldkey_hotkey_identities(),
-            subtensor.get_delegate_identities(),
-            subtensor.all_subnets(),
-            subtensor.get_balance(wallet.coldkeypub.ss58_address),
-        )
-
-        if unstake_all_alpha:
-            stake_info = [stake for stake in stake_info if stake.netuid != 0]
-
-        if not stake_info:
-            console.print("[red]No stakes found to unstake[/red]")
-            return False
-
-        all_sn_dynamic_info = {info.netuid: info for info in all_sn_dynamic_info_}
-
-        # Calculate total value and slippage for all stakes
-        total_received_value = Balance(0)
-        table_title = (
-            "Unstaking Summary - All Stakes"
-            if not unstake_all_alpha
-            else "Unstaking Summary - All Alpha Stakes"
-        )
-        table = Table(
-            title=(
-                f"\n[{COLOR_PALETTE['GENERAL']['HEADER']}]{table_title}[/{COLOR_PALETTE['GENERAL']['HEADER']}]\n"
-                f"Wallet: [{COLOR_PALETTE['GENERAL']['COLDKEY']}]{wallet.name}[/{COLOR_PALETTE['GENERAL']['COLDKEY']}], "
-                f"Coldkey ss58: [{COLOR_PALETTE['GENERAL']['COLDKEY']}]{wallet.coldkeypub.ss58_address}[/{COLOR_PALETTE['GENERAL']['COLDKEY']}]\n"
-                f"Network: [{COLOR_PALETTE['GENERAL']['HEADER']}]{subtensor.network}[/{COLOR_PALETTE['GENERAL']['HEADER']}]\n"
-            ),
-            show_footer=True,
-            show_edge=False,
-            header_style="bold white",
-            border_style="bright_black",
-            style="bold",
-            title_justify="center",
-            show_lines=False,
-            pad_edge=True,
-        )
-        table.add_column("Netuid", justify="center", style="grey89")
-        table.add_column(
-            "Hotkey", justify="center", style=COLOR_PALETTE["GENERAL"]["HOTKEY"]
-        )
-        table.add_column(
-            f"Current Stake ({Balance.get_unit(1)})",
-            justify="center",
-            style=COLOR_PALETTE["STAKE"]["STAKE_ALPHA"],
-        )
-        table.add_column(
-            f"Rate ({Balance.unit}/{Balance.get_unit(1)})",
-            justify="center",
-            style=COLOR_PALETTE["POOLS"]["RATE"],
-        )
-        table.add_column(
-            f"Recieved ({Balance.unit})",
-            justify="center",
-            style=COLOR_PALETTE["POOLS"]["TAO_EQUIV"],
-        )
-        table.add_column(
-            "Slippage",
-            justify="center",
-            style=COLOR_PALETTE["STAKE"]["SLIPPAGE_PERCENT"],
-        )
-        max_slippage = 0.0
-        for stake in stake_info:
-            if stake.stake.rao == 0:
-                continue
-
-            dynamic_info = all_sn_dynamic_info.get(stake.netuid)
-            stake_amount = stake.stake
-            received_amount, _, slippage_pct_float = (
-                dynamic_info.alpha_to_tao_with_slippage(stake_amount)
-            )
-
-            total_received_value += received_amount
-
-            # Get hotkey identity
-            if hk_identity := ck_hk_identities["hotkeys"].get(stake.hotkey_ss58):
-                hotkey_name = hk_identity.get("identity", {}).get(
-                    "name", ""
-                ) or hk_identity.get("display", "~")
-                hotkey_display = f"{hotkey_name}"
-            elif old_identity := old_identities.get(stake.hotkey_ss58):
-                hotkey_name = old_identity.display
-                hotkey_display = f"{hotkey_name}"
-            else:
-                hotkey_display = stake.hotkey_ss58
-
-            if dynamic_info.is_dynamic:
-                slippage_pct = f"{slippage_pct_float:.4f} %"
-            else:
-                slippage_pct_float = 0
-                slippage_pct = "[red]N/A[/red]"
-
-            max_slippage = max(max_slippage, slippage_pct_float)
-
-            table.add_row(
-                str(stake.netuid),
-                hotkey_display,
-                str(stake_amount),
-                str(float(dynamic_info.price))
-                + f"({Balance.get_unit(0)}/{Balance.get_unit(stake.netuid)})",
-                str(received_amount),
-                slippage_pct,
-            )
-    console.print(table)
-    message = ""
-    if max_slippage > 5:
-        message += f"[{COLOR_PALETTE['STAKE']['SLIPPAGE_TEXT']}]-------------------------------------------------------------------------------------------------------------------\n"
-        message += f"[bold]WARNING:[/bold] The slippage on one of your operations is high: [{COLOR_PALETTE['STAKE']['SLIPPAGE_PERCENT']}]{max_slippage:.4f}%[/{COLOR_PALETTE['STAKE']['SLIPPAGE_PERCENT']}], this may result in a loss of funds.\n"
-        message += "-------------------------------------------------------------------------------------------------------------------\n"
-        console.print(message)
-
-    console.print(
-        f"Expected return after slippage: [{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]{total_received_value}"
-    )
-
-    if prompt and not Confirm.ask(
-        "\nDo you want to proceed with unstaking everything?"
-    ):
-        return False
-
-    try:
-        wallet.unlock_coldkey()
-    except KeyFileError:
-        err_console.print("Error decrypting coldkey (possibly incorrect password)")
-        return False
-
-    console_status = (
-        ":satellite: Unstaking all Alpha stakes..."
-        if unstake_all_alpha
-        else ":satellite: Unstaking all stakes..."
-    )
-    with console.status(console_status):
-        call_function = "unstake_all_alpha" if unstake_all_alpha else "unstake_all"
-        call = await subtensor.substrate.compose_call(
-            call_module="SubtensorModule",
-            call_function=call_function,
-            call_params={"hotkey": wallet.hotkey.ss58_address},
-        )
-        success, error_message = await subtensor.sign_and_send_extrinsic(
-            call=call,
-            wallet=wallet,
-            wait_for_inclusion=True,
-            wait_for_finalization=False,
-        )
-
-        if success:
-            success_message = (
-                ":white_heavy_check_mark: [green]Successfully unstaked all stakes[/green]"
-                if not unstake_all_alpha
-                else ":white_heavy_check_mark: [green]Successfully unstaked all Alpha stakes[/green]"
-            )
-            console.print(success_message)
-            new_balance = await subtensor.get_balance(wallet.coldkeypub.ss58_address)
-            console.print(
-                f"Balance:\n [blue]{current_wallet_balance}[/blue] :arrow_right: [{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]{new_balance}"
-            )
-            return True
-        else:
-            err_console.print(
-                f":cross_mark: [red]Failed to unstake[/red]: {error_message}"
-            )
-            return False
-
-
-async def unstake(
-    wallet: Wallet,
-    subtensor: "SubtensorInterface",
-    hotkey_ss58_address: str,
-    all_hotkeys: bool,
-    include_hotkeys: list[str],
-    exclude_hotkeys: list[str],
-    amount: float,
-    keep_stake: float,
-    unstake_all: bool,
-    prompt: bool,
-    interactive: bool = False,
-    netuid: Optional[int] = None,
-    unstake_all_alpha: bool = False,
-):
-    """Unstake tokens from hotkey(s)."""
-
-    if unstake_all or unstake_all_alpha:
-        return await _unstake_all(wallet, subtensor, unstake_all_alpha, prompt)
-
-    unstake_all_from_hk = False
-    with console.status(
-        f"Retrieving subnet data & identities from {subtensor.network}...",
-        spinner="earth",
-    ):
-        all_sn_dynamic_info_, ck_hk_identities, old_identities = await asyncio.gather(
-            subtensor.all_subnets(),
-            subtensor.fetch_coldkey_hotkey_identities(),
-            subtensor.get_delegate_identities(),
-        )
-        all_sn_dynamic_info = {info.netuid: info for info in all_sn_dynamic_info_}
-
-    if interactive:
-        hotkeys_to_unstake_from, unstake_all_from_hk = await unstake_selection(
-            subtensor,
-            wallet,
-            all_sn_dynamic_info,
-            ck_hk_identities,
-            old_identities,
-            netuid=netuid,
-        )
-        if not hotkeys_to_unstake_from:
-            console.print("[red]No unstake operations to perform.[/red]")
-            return False
-        netuids = list({netuid for _, _, netuid in hotkeys_to_unstake_from})
-
-    else:
-        netuids = (
-            [int(netuid)]
-            if netuid is not None
-            else await subtensor.get_all_subnet_netuids()
-        )
-
-        # Get the hotkey_names (if any) and the hotkey_ss58s.
-        hotkeys_to_unstake_from: list[tuple[Optional[str], str]] = []
-        if hotkey_ss58_address:
-            print_verbose(f"Unstaking from ss58 ({hotkey_ss58_address})")
-            # Unstake from specific hotkey.
-            hotkeys_to_unstake_from = [(None, hotkey_ss58_address)]
-        elif all_hotkeys:
-            print_verbose("Unstaking from all hotkeys")
-            # Unstake from all hotkeys.
-            all_hotkeys_: list[Wallet] = get_hotkey_wallets_for_wallet(wallet=wallet)
-            # Exclude hotkeys that are specified.
-            hotkeys_to_unstake_from = [
-                (wallet.hotkey_str, wallet.hotkey.ss58_address)
-                for wallet in all_hotkeys_
-                if wallet.hotkey_str not in exclude_hotkeys
-            ]
-        elif include_hotkeys:
-            print_verbose("Unstaking from included hotkeys")
-            # Unstake from specific hotkeys.
-            for hotkey_identifier in include_hotkeys:
-                if is_valid_ss58_address(hotkey_identifier):
-                    # If the hotkey is a valid ss58 address, we add it to the list.
-                    hotkeys_to_unstake_from.append((None, hotkey_identifier))
-                else:
-                    # If the hotkey is not a valid ss58 address, we assume it is a hotkey name.
-                    # We then get the hotkey from the wallet and add it to the list.
-                    wallet_ = Wallet(
-                        name=wallet.name,
-                        path=wallet.path,
-                        hotkey=hotkey_identifier,
-                    )
-                    hotkeys_to_unstake_from.append(
-                        (wallet_.hotkey_str, wallet_.hotkey.ss58_address)
-                    )
-        else:
-            # Only cli.config.wallet.hotkey is specified.
-            # So we unstake from that single hotkey.
-            print_verbose(
-                f"Unstaking from wallet: ({wallet.name}) from hotkey: ({wallet.hotkey_str})"
-            )
-            assert wallet.hotkey is not None
-            hotkeys_to_unstake_from = [(wallet.hotkey_str, wallet.hotkey.ss58_address)]
-
-    with console.status(
-        f"Retrieving stake data from {subtensor.network}...",
-        spinner="earth",
-    ):
-        # Prepare unstaking transactions
-        unstake_operations = []
-        total_received_amount = Balance.from_tao(0)
-        current_wallet_balance: Balance = await subtensor.get_balance(
-            wallet.coldkeypub.ss58_address
-        )
-        max_float_slippage = 0
-
-        # Fetch stake balances
-        chain_head = await subtensor.substrate.get_chain_head()
-        stake_info_list = await subtensor.get_stake_for_coldkey(
-            coldkey_ss58=wallet.coldkeypub.ss58_address,
-            block_hash=chain_head,
-        )
-        stake_in_netuids = {}
-        for stake_info in stake_info_list:
-            if stake_info.hotkey_ss58 not in stake_in_netuids:
-                stake_in_netuids[stake_info.hotkey_ss58] = {}
-            stake_in_netuids[stake_info.hotkey_ss58][stake_info.netuid] = (
-                stake_info.stake
-            )
-
-    # Flag to check if user wants to quit
-    skip_remaining_subnets = False
-    if hotkeys_to_unstake_from:
-        console.print(
-            "[dark_sea_green3]Tip: Enter 'q' any time to skip further entries and process existing unstakes"
-        )
-
-    # Iterate over hotkeys and netuids to collect unstake operations
-    unstake_all_hk_ss58 = None
-    for hotkey in hotkeys_to_unstake_from:
-        if skip_remaining_subnets:
-            break
-
-        if interactive:
-            staking_address_name, staking_address_ss58, netuid = hotkey
-            netuids_to_process = [netuid]
-        else:
-            staking_address_name, staking_address_ss58 = hotkey
-            netuids_to_process = netuids
-
-        initial_amount = amount
-
-        if len(netuids_to_process) > 1:
-            console.print(
-                "[dark_sea_green3]Tip: Enter 'q' any time to stop going over remaining subnets and process current unstakes.\n"
-            )
-
-        for netuid in netuids_to_process:
-            if skip_remaining_subnets:
-                break  # Exit the loop over netuids
-
-            dynamic_info = all_sn_dynamic_info.get(netuid)
-            current_stake_balance = stake_in_netuids[staking_address_ss58][netuid]
-            if current_stake_balance.tao == 0:
-                continue  # No stake to unstake
-
-            # Determine the amount we are unstaking.
-            if unstake_all_from_hk or unstake_all:
-                amount_to_unstake_as_balance = current_stake_balance
-                unstake_all_hk_ss58 = staking_address_ss58
-            elif initial_amount:
-                amount_to_unstake_as_balance = Balance.from_tao(initial_amount)
-            else:
-                amount_to_unstake_as_balance = ask_unstake_amount(
-                    current_stake_balance,
-                    netuid,
-                    staking_address_name
-                    if staking_address_name
-                    else staking_address_ss58,
-                    staking_address_ss58,
-                    interactive,
-                )
-                if amount_to_unstake_as_balance is None:
-                    skip_remaining_subnets = True
-                    break
-
-            # Check enough stake to remove.
-            amount_to_unstake_as_balance.set_unit(netuid)
-            if amount_to_unstake_as_balance > current_stake_balance:
-                err_console.print(
-                    f"[red]Not enough stake to remove[/red]:\n Stake balance: [dark_orange]{current_stake_balance}[/dark_orange]"
-                    f" < Unstaking amount: [dark_orange]{amount_to_unstake_as_balance}[/dark_orange]"
-                )
-                continue  # Skip to the next subnet - useful when single amount is specified for all subnets
-
-            received_amount, _, slippage_pct_float = (
-                dynamic_info.alpha_to_tao_with_slippage(amount_to_unstake_as_balance)
-            )
-            total_received_amount += received_amount
-            if dynamic_info.is_dynamic:
-                slippage_pct = f"{slippage_pct_float:.4f} %"
-            else:
-                slippage_pct_float = 0
-                slippage_pct = "[red]N/A[/red]"
-            max_float_slippage = max(max_float_slippage, slippage_pct_float)
-
-            unstake_operations.append(
-                {
-                    "netuid": netuid,
-                    "hotkey_name": staking_address_name
-                    if staking_address_name
-                    else staking_address_ss58,
-                    "hotkey_ss58": staking_address_ss58,
-                    "amount_to_unstake": amount_to_unstake_as_balance,
-                    "current_stake_balance": current_stake_balance,
-                    "received_amount": received_amount,
-                    "slippage_pct": slippage_pct,
-                    "slippage_pct_float": slippage_pct_float,
-                    "dynamic_info": dynamic_info,
-                }
-            )
-
-    if not unstake_operations:
-        console.print("[red]No unstake operations to perform.[/red]")
-        return False
-
-    # Build the table
-    table = Table(
-        title=f"\n[{COLOR_PALETTE['GENERAL']['HEADER']}]Unstaking to: \nWallet: [{COLOR_PALETTE['GENERAL']['COLDKEY']}]{wallet.name}[/{COLOR_PALETTE['GENERAL']['COLDKEY']}],"
-        f" Coldkey ss58: [{COLOR_PALETTE['GENERAL']['COLDKEY']}]{wallet.coldkeypub.ss58_address}[/{COLOR_PALETTE['GENERAL']['COLDKEY']}]\n"
-        f"Network: {subtensor.network}[/{COLOR_PALETTE['GENERAL']['HEADER']}]\n",
-        show_footer=True,
-        show_edge=False,
-        header_style="bold white",
-        border_style="bright_black",
-        style="bold",
-        title_justify="center",
-        show_lines=False,
-        pad_edge=True,
-    )
-    table.add_column("Netuid", justify="center", style="grey89")
-    table.add_column(
-        "Hotkey", justify="center", style=COLOR_PALETTE["GENERAL"]["HOTKEY"]
-    )
-    table.add_column(
-        f"Amount ({Balance.get_unit(1)})",
-        justify="center",
-        style=COLOR_PALETTE["POOLS"]["TAO"],
-    )
-    table.add_column(
-        f"Rate ({Balance.get_unit(0)}/{Balance.get_unit(1)})",
-        justify="center",
-        style=COLOR_PALETTE["POOLS"]["RATE"],
-    )
-    table.add_column(
-        f"Received ({Balance.get_unit(0)})",
-        justify="center",
-        style=COLOR_PALETTE["POOLS"]["TAO_EQUIV"],
-        footer=f"{total_received_amount}",
-    )
-    table.add_column(
-        "Slippage", justify="center", style=COLOR_PALETTE["STAKE"]["SLIPPAGE_PERCENT"]
-    )
-
-    for op in unstake_operations:
-        dynamic_info = op["dynamic_info"]
-        table.add_row(
-            str(op["netuid"]),
-            op["hotkey_name"],
-            str(op["amount_to_unstake"]),
-            str(float(dynamic_info.price))
-            + f"({Balance.get_unit(0)}/{Balance.get_unit(op['netuid'])})",
-            str(op["received_amount"]),
-            op["slippage_pct"],
-        )
-
-    console.print(table)
-
-    if max_float_slippage > 5:
-        console.print(
-            "\n"
-            f"[{COLOR_PALETTE['STAKE']['SLIPPAGE_TEXT']}]-------------------------------------------------------------------------------------------------------------------\n"
-            f"[bold]WARNING:[/bold]  The slippage on one of your operations is high: [{COLOR_PALETTE['STAKE']['SLIPPAGE_PERCENT']}]{max_float_slippage} %[/{COLOR_PALETTE['STAKE']['SLIPPAGE_PERCENT']}],"
-            " this may result in a loss of funds.\n"
-            f"-------------------------------------------------------------------------------------------------------------------\n"
-        )
-
-    console.print(
-        """
-[bold white]Description[/bold white]:
-The table displays information about the stake remove operation you are about to perform.
-The columns are as follows:
-    - [bold white]Netuid[/bold white]: The netuid of the subnet you are unstaking from.
-    - [bold white]Hotkey[/bold white]: The ss58 address or identity of the hotkey you are unstaking from. 
-    - [bold white]Amount[/bold white]: The stake amount you are removing from this key.
-    - [bold white]Rate[/bold white]: The rate of exchange between TAO and the subnet's stake.
-    - [bold white]Received[/bold white]: The amount of free balance TAO you will receive on this subnet after slippage.
-    - [bold white]Slippage[/bold white]: The slippage percentage of the unstake operation. (0% if the subnet is not dynamic i.e. root).
-"""
-    )
-    if prompt:
-        if not Confirm.ask("Would you like to continue?"):
-            raise typer.Exit()
-
-    # Perform unstaking operations
-    try:
-        wallet.unlock_coldkey()
-    except KeyFileError:
-        err_console.print("Error decrypting coldkey (possibly incorrect password)")
-        return False
-
-    with console.status("\n:satellite: Performing unstaking operations...") as status:
-        if unstake_all_from_hk:
-            call = await subtensor.substrate.compose_call(
-                call_module="SubtensorModule",
-                call_function="unstake_all",
-                call_params={
-                    "hotkey": unstake_all_hk_ss58,
-                },
-            )
-            extrinsic = await subtensor.substrate.create_signed_extrinsic(
-                call=call, keypair=wallet.coldkey
-            )
-            response = await subtensor.substrate.submit_extrinsic(
-                extrinsic, wait_for_inclusion=True, wait_for_finalization=False
-            )
-            await response.process_events()
-            if not await response.is_success:
-                print_error(
-                    f":cross_mark: [red]Failed[/red] with error: "
-                    f"{format_error_message(await response.error_message, subtensor.substrate)}",
-                    status,
-                )
-            else:
-                new_balance = await subtensor.get_balance(
-                    wallet.coldkeypub.ss58_address
-                )
-                console.print(":white_heavy_check_mark: [green]Finalized[/green]")
-                console.print(
-                    f"Balance:\n  [blue]{current_wallet_balance}[/blue] :arrow_right: [{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]{new_balance}"
-                )
-        else:
-            for op in unstake_operations:
-                netuid_i = op["netuid"]
-                staking_address_name = op["hotkey_name"]
-                staking_address_ss58 = op["hotkey_ss58"]
-                amount = op["amount_to_unstake"]
-                current_stake_balance = op["current_stake_balance"]
-
-                status.update(
-                    f"\n:satellite: Unstaking {amount} from {staking_address_name} on netuid: {netuid_i} ..."
-                )
-
-                call = await subtensor.substrate.compose_call(
-                    call_module="SubtensorModule",
-                    call_function="remove_stake",
-                    call_params={
-                        "hotkey": staking_address_ss58,
-                        "netuid": netuid_i,
-                        "amount_unstaked": amount.rao,
-                    },
-                )
-                extrinsic = await subtensor.substrate.create_signed_extrinsic(
-                    call=call, keypair=wallet.coldkey
-                )
-                response = await subtensor.substrate.submit_extrinsic(
-                    extrinsic, wait_for_inclusion=True, wait_for_finalization=False
-                )
-                await response.process_events()
-                if not await response.is_success:
-                    print_error(
-                        f":cross_mark: [red]Failed[/red] with error: "
-                        f"{format_error_message(await response.error_message, subtensor.substrate)}",
-                        status,
-                    )
-                else:
-                    new_balance = await subtensor.get_balance(
-                        wallet.coldkeypub.ss58_address
-                    )
-                    new_stake_info = await subtensor.get_stake_for_coldkey(
-                        coldkey_ss58=wallet.coldkeypub.ss58_address,
-                    )
-                    new_stake = Balance.from_rao(0)
-                    for stake_info in new_stake_info:
-                        if (
-                            stake_info.hotkey_ss58 == staking_address_ss58
-                            and stake_info.netuid == netuid_i
-                        ):
-                            new_stake = stake_info.stake.set_unit(netuid_i)
-                            break
-                    console.print(":white_heavy_check_mark: [green]Finalized[/green]")
-                    console.print(
-                        f"Balance:\n  [blue]{current_wallet_balance}[/blue] :arrow_right: [{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]{new_balance}"
-                    )
-                    console.print(
-                        f"Subnet: [{COLOR_PALETTE['GENERAL']['SUBHEADING']}]{netuid_i}[/{COLOR_PALETTE['GENERAL']['SUBHEADING']}]"
-                        f" Stake:\n  [blue]{current_stake_balance}[/blue] :arrow_right: [{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]{new_stake}"
-                    )
-    console.print(
-        f"[{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]Unstaking operations completed."
-    )
 
 
 async def stake_list(
